feat: add sticky header to Projects page
Drop/revert this commit if the sticky header is not desired.
This also removes `TempLayout`, since it's only used here.
